Refactor: standardize React component filenames to PascalCase
